requirements.txt
flask
flask-session
requests
redis
python-dotenv